+++ src/lib/demo/layout/LayoutHeader.svelte
@@ -3,7 +3,7 @@
 	import { onMount } from 'svelte';
 	import Logo from '$lib/demo/Logo.svelte';
 
-	export let menu: string = '/';
+	export let menu: string = '/home';
 	let active: string = 'light';
 	onMount(() => {
 		const _theme = getStorage('BEER_THEME');
@@ -24,7 +24,7 @@
 	</div>
 	<div class='nav'>
 		<a class:active={menu === '/'} href='/'>主页</a>
-		<a class:active={menu === '/svg'} href='/components'>组件</a>
+		<a class:active={menu === '/components'} href='/components'>组件</a>
 		<a target='_blank' href='https://github.com/beerui/BeerUi/blob/main/CHANGELOG.md'>更新日志</a>
 		<a class:active={menu === '/others'} href='/others'>其它</a>
 		<div class='tabs'>