+++ svelte.config.js
@@ -1,4 +1,5 @@
 import preprocess from 'svelte-preprocess';
+import adapter from '@sveltejs/adapter-static';
 import vercel from '@sveltejs/adapter-vercel';
 // 如果需要打包发布到 vercel 打开注释即可
 const config = {
@@ -31,7 +32,14 @@ const config = {
 	],
 	kit: {
 		// hydrate the <div id="svelte"> element in src/app.html
-		adapter: vercel()
+		adapter: process.env.NODE_ENV !== 'office' ? vercel() : adapter({
+			// default options are shown. On some platforms
+			// these options are set automatically — see below
+			pages: 'build',
+			assets: 'build',
+			precompress: false,
+			strict: true
+		})
 	},
 };